Changed Grid Design
The semicircles walls in the previous design were added in order to provide stability for the walls and to make the grid less flexible.
After printing it though. I realized the grid was already relatively stiff and didn't really need the additional stiffness. Furthermore, it made it harder to determine what was and wasn't a wall.

The new design is much simpler and completely removes the semicircle walls on the grid. Additionally, it has fillets and chamfers in order to reduce the effects of print artifacts such as elephants foot.