Changed the turtle simulation code to better match what how the robot would move
Changed the grid size and start and end positions to match the maze the robot will solve.
Also changed the movement of the simulated robot to move to the walls it checks to better visualize how the robot will move and make it easier to see how future optimizations would affect the robot.