Added basic functions for movement
Added Functions for moving the motors as well as getting the x and y position in mm.

The movement functions for the x and y axis make each axis move at the same speed if the same speed value is passed in despite different gear ratios.
EX: Even if running the x axis motors at max speed normally moves the axis faster than running the y axis motor at top speed, when running it through the movex() and movey() functions, the top speed of the x axis will be slowed in order to match that of the y axis.