A test version that tries to use positional information to detect when the robot has collided with something
This test version is an attempt to try to calculate the change in position over a short interval in order to try to determine when the robot has collided with something.
This switch was made because the current detection isn't ideal because the current changes depending on the battery voltage.
This currently doesn't work for reasons unknown. Furthermore, this method isn't great either because it depends on the fact that the barrier is strong enough to stop the robot. If the barrier isn't strong enough, it could break and the robot would not detect it.