Smaller maze and more rigid mount
The old maze was too big and the device could not reach each square. So I reduced the size of the grid.

In the old version, the maze mounts were independent and could flex a bit.
This time, the mounts are still separate objects but there are connectors that connect them. One connector uses a dovetail joint while the other has a different kind of joint. The reason for the different joints is so that no support material is needed when printing.