Added more base movement functions
This branch is focused on creating the base movement functions of the robot.

Added Functions:
collideX() - detects collision in X axis
collideY() - detects collision in Y axis

homeDevice() - finds the lower left corner of the device bounds and zeros it there.

goto(x,y) - A function that moves the robot to a specified position in millimeters
gotoCol(x,y) - A variation of the goto function that is also supposed to detect collisions and react accordingly.

zeroMaze() - A function that moves the robot to the (0,0) grid position and zeros the robot there.

moveVerticalTile(tiles)
moveHorizontalTile(tiles)
movement function that will move the robot a specified number of grid tiles in either the x or y direction.

This version of code is currently untested and requires testing to determine certain values. For example, the position of the (0,0) grid tile is unknown.