Failed version of gotoCol()
The function gotoCol() is supposed to detect if the robot collides with something when moving to a specific point. However, the maze walls I designed are not strong enough and do not make the torque-based collision detection go off. Instead, the robot just plows through the wall breaking it.
Lowering the threshold for the collision detection does not work as it simply causes false collisions when moving along the track