Created IR Sensor Mount and Improved Homing Sequence
Since the torque based wall detection didn't work, I decided to switch to IR Sensors for wall detection. I plan to have the sensors face upwards and try to detect the gaps in the wall holder. If there is a wall,  then the gap will be filled. But if there isn't, there will be a gap that the IR sensor can detect.
Since these gaps are very small, I improved the homing sequence by adding a second homing at a slower speed for a more accurate reading.